Change all time parameters from float to double to prevent drift
Unity stores Time.time as a float, which accumulates floating-point
drift after ~7 days of uptime. Changed all time-related parameters,
fields, and properties across the entire core library from float to
double. Use Time.timeAsDouble in Unity integration.

Affected: Machine, Scheduler, MachineContext, StepResult, WaitForTimeStep,
TransitionTrace, DebugSnapshot, FsmDebugController, MachineRecovery,
CompositeStateMachine, StateBuilder. All samples and docs updated.

+++ samples/PickGame/Program.cs
@@ -285,7 +285,7 @@ class Program
         Console.ResetColor();
     }
 
-    static void PrintTrace(string from, string to, TransitionReasonKind reason, string detail, float time)
+    static void PrintTrace(string from, string to, TransitionReasonKind reason, string detail, double time)
     {
         Console.ForegroundColor = ConsoleColor.DarkYellow;
         Console.WriteLine($"  ── {from} → {to} | {reason}: {detail} | t={time:F1}s");
